[M1_TR-204] Moved search directory under /api
