+++ server/src/api/jobs/jobs.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { PrismaService } from '../../database/connection.service';
-import { SearchModule } from '../../search/search.module';
+import { SearchModule } from '../search/search.module';
 import { JobsController } from './jobs.controller';
 import { JobsService } from './jobs.service';
