+++ server/src/api/search/search.module.ts
