+++ server/src/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { JobsModule } from './api/jobs/jobs.module';
+import { SearchModule } from './api/search/search.module';
 import { DatabaseModule } from './database/database.module';
-import { SearchModule } from './search/search.module';
 
 @Module({
   imports: [DatabaseModule, JobsModule, SearchModule],
